Add tweaker loader (#7)
This PR allows mod developers to bundle the mod api into their mod
without causing version conflicts. To do so it employs a tweaker,
something already used by a few libraries in the hypixel legacy modding
scene, so this should be a fairly familiar system to a lot of devs. Note
that this one loads from a bundled JAR instead of a network connection.

---------

Co-authored-by: ILikePlayingGames <22475143+ILikePlayingGames@users.noreply.github.com>
Co-authored-by: Connor Linfoot <connorlinfoot@me.com>